% Generated by build script
\name{nyfed_effr}
\alias{nyfed_effr}
\title{Fetch EFFR (Effective Federal Funds Rate) data}
\usage{
nyfed_effr()
}
\arguments{
\item{start}{Start date (Date or "YYYY-MM-DD"), default last 30 days}
\item{end}{End date (Date or "YYYY-MM-DD"), default today}
}
\value{tibble: date, rate, percentile_1, percentile_25,}
\description{Fetch EFFR (Effective Federal Funds Rate) data}
