% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyfed.R
\name{nyfed_effr}
\alias{nyfed_effr}
\title{Fetch EFFR (Effective Federal Funds Rate) data}
\usage{
nyfed_effr(start = NULL, end = NULL)
}
\arguments{
\item{start}{Start date (Date or "YYYY-MM-DD"), default last 30 days}

\item{end}{End date (Date or "YYYY-MM-DD"), default today}
}
\value{
tibble: date, rate, percentile_1, percentile_25,
percentile_75, percentile_99, volume
}
\description{
Fetch EFFR (Effective Federal Funds Rate) data
}
