% Generated by build script
\name{nyfed_sofr}
\alias{nyfed_sofr}
\title{Fetch SOFR (Secured Overnight Financing Rate) data}
\usage{
nyfed_sofr()
}
\arguments{
\item{start}{Start date (Date or "YYYY-MM-DD"), default last 30 days}
\item{end}{End date (Date or "YYYY-MM-DD"), default today}
}
\value{tibble: date, rate, percentile_1, percentile_25,}
\description{Fetch SOFR (Secured Overnight Financing Rate) data}
